requirements.txt
torch>=2.0.1
pandas>=2.2.2
sacremoses>=0.1.1
transformers>=4.48.3
numpy>=1.26.4
tqdm>=4.66.4
requests-toolbelt==0.10.1
urllib3==1.26.15
sentencepiece>=0.2.0
protobuf>=5.27.1
matplotlib>=3.9.2
scikit-learn>=1.5.2
umap-learn>=0.5.6
plotly>=5.24.1
kaleido!=0.2.1.post1
dask[complete]>=2024.11.2
pyarrow>=18.1.0
einops>=0.8.0
datasets>=3.1.0